refactor: simplify command, handle '=' symbol,  misc command output updates
